requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.44.0
Django==5.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-jazzmin==3.0.1
djangorestframework==3.16.0
idna==3.10
psycopg2-binary==2.9.10

requests==2.32.3
six==1.17.0
sqlparse==0.5.3
urllib3==2.4.0
